server/requirements.txt
strawberry-graphql==0.113.0
flask[async]==2.1.2
gunicorn==20.1.0
